Found the bug I had in the time measurements
I had reverted he time change since I was getting an out of bound
exception, but now it works well!!! So every measurement (or average of
3 measurements) is about 120ms apart, and it's very consistant. After
about 4 measurements it adds another 1 ms for the calculations etc, but
that seems small. And it's easy to add arbitrary time delays so that
measurements are further apart. The only thing now is, it's messed up
the graphing portion of the gui, since the last byte is now the time, I
might fix that in a sec. All I need to do is reverse it I think.