My rewrite
Completely rewrote the program to make it super simple so it's easier
for me to debug. Made the data file just save in the same directory, and
ask for the filename at the start of the program, before anything else
happens. Got rid of the GUI class and we'll see what happens.