Made input name a variable
The input method for getFileName didn't work, so I just made it a
variable, you'll have to change it each time you want a new file
otherwise it'll overwrite what you already made. I know it's not ideal,
but it works and is simple. I'll change it if I have time. The text file
is an example of data I got while NOT at the spectrophotometer.